Added is game over check and screen to bejeweled

+++ puzzlebobble/src/main/java/edu/uci/puzzlebobble/PuzzleBobbleViewController.java
@@ -103,7 +103,7 @@ public class PuzzleBobbleViewController extends StackPane implements Pausable {
                     if (board.isGameOver()) {
                         gameOverScreen.setVisible(true);
                         tilePane.setDisable(true);
-                        finalScore.setText("Final Score : " + board.getScore());
+                        finalScore.setText("Final Score : " + (int) board.getScore());
                         gameOver.set(true);
                     }
